+++ lib/src/helpers/date_time_extensions.dart
@@ -1,0 +1,5 @@
+extension DateTimeExtensions on DateTime {
+  bool isSameDay({required DateTime dateTime}){
+    return year == dateTime.year && month == dateTime.month && day == dateTime.day;
+  }
+}